[IMP] procurement_plan: keep procurement_method_id in purchase_request_budget_procurement
Fold the procurement_plan_purchase bridge into purchase_request_budget_procurement
instead of shipping a separate one-field module. procurement.plan.procurement_method_id
(M2O procurement.method) plus its tree/form/search views and the action_ready check now
live in purchase_request_budget_procurement, which already integrates purchase requests
with procurement plans.

Base procurement_plan stays decoupled from purchase_request_kmitl (depends only on web,
budget); purchase_request_eta remains in base. procurement_plan_portal and
procurement_plan_demo now depend on purchase_request_budget_procurement for the field.
