+++ procurement_plan_portal/__manifest__.py
@@ -5,7 +5,7 @@
     "author": "",
     "website": "",
     "category": "",
-    "depends": ["portal", "procurement_plan", "procurement_plan_purchase"],
+    "depends": ["portal", "procurement_plan", "purchase_request_budget_procurement"],
     "data": ["views/templates.xml"],
     "assets": {
         "web.assets_frontend": [
