--- procurement_plan_purchase/__manifest__.py
@@ -1,15 +1,0 @@
-{
-    "name": "Procurement Plan Purchase",
-    "version": "16.0.1.0.0",
-    "author": "Aginix Technologies",
-    "website": "https://github.com/aginix/kmitl",
-    "category": "KMITL",
-    "depends": ["procurement_plan", "purchase_request_kmitl"],
-    "data": [
-        "views/procurement_plan_views.xml",
-    ],
-    "application": False,
-    "installable": True,
-    "auto_install": False,
-    "license": "LGPL-3",
-}
