--- procurement_plan_purchase/i18n/th.po
@@ -1,20 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* procurement_plan_purchase
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"Language: th\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-
-#. module: procurement_plan_purchase
-#: model:ir.model.fields,field_description:procurement_plan_purchase.field_procurement_plan__procurement_method_id
-msgid "Procurement Method"
-msgstr "วิธีการจัดซื้อจัดจ้าง"
