--- procurement_plan_purchase/models/__init__.py
@@ -1,1 +1,0 @@
-from . import procurement_plan
