--- procurement_plan_purchase/views/procurement_plan_views.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-
-    <!-- Tree: add procurement method column -->
-    <record id="view_procurement_plan_tree_purchase" model="ir.ui.view">
-        <field name="name">view.procurement.plan.tree.purchase</field>
-        <field name="model">procurement.plan</field>
-        <field name="inherit_id" ref="procurement_plan.view_procurement_plan_tree" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='activity_analytic_id']" position="after">
-                <field name="procurement_method_id" optional="show" />
-            </xpath>
-        </field>
-    </record>
-
-    <!-- Form: add procurement method before the purchase request ETA -->
-    <record id="view_procurement_plan_form_purchase" model="ir.ui.view">
-        <field name="name">view.procurement.plan.form.purchase</field>
-        <field name="model">procurement.plan</field>
-        <field name="inherit_id" ref="procurement_plan.view_procurement_plan_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='purchase_request_eta']" position="before">
-                <field name="procurement_method_id" widget="radio" attrs="{'readonly': [('state', 'not in', ('draft', 'new'))]}" />
-            </xpath>
-        </field>
-    </record>
-
-    <!-- Search: add procurement method field, group by and search panel -->
-    <record id="view_procurement_plan_search_purchase" model="ir.ui.view">
-        <field name="name">view.procurement.plan.search.purchase</field>
-        <field name="model">procurement.plan</field>
-        <field name="inherit_id" ref="procurement_plan.view_procurement_plan_search" />
-        <field name="arch" type="xml">
-            <xpath expr="//search/field[@name='activity_analytic_id']" position="after">
-                <field name="procurement_method_id" />
-            </xpath>
-            <xpath expr="//filter[@name='group_by_state']" position="before">
-                <filter string="วิธีการจัดซื้อจัดจ้าง" name="group_by_procurement_method" context="{'group_by': 'procurement_method_id'}" />
-            </xpath>
-            <xpath expr="//searchpanel/field[@name='fund_analytic_id']" position="after">
-                <field name="procurement_method_id" icon="fa-calendar" select="multi" />
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
