+++ purchase_request_budget_procurement/models/purchase_request.py
@@ -217,6 +217,19 @@ class PurchaseRequest(models.Model):
 class ProcurementPlan(models.Model):
     _inherit = "procurement.plan"
 
+    procurement_method_id = fields.Many2one(
+        comodel_name="procurement.method",
+        string="Procurement Method",
+        required=False,
+        tracking=True,
+    )
+
+    def action_ready(self):
+        for rec in self:
+            if not rec.procurement_method_id:
+                raise UserError(_("กรุณาระบุแผนการดำเนินงานให้เสร็จสิ้นทั้งหมด"))
+        return super().action_ready()
+
     purchase_request_count = fields.Integer(
         string="Purchase Requests Count", compute="_compute_purchase_request_count"
     )
